requirements.txt
asgiref==3.11.1
asttokens==3.0.1
colorama==0.4.6
decorator==5.2.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
python-decouple==3.8
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.6.0
